patch version number in setup.py
The version number in the setup.py file does not match the version of the
source code.  This leads to a mismatch between the package version and the one
recorded in the .egg-info directory, Keras_Applications-1.0.2-py2.7.egg-info/
which should be Keras_Applications-1.0.4-py2.7.egg-info/

The included patch correct the version number in the setup.py file.